perf(core): save route history for horizontal tabs
