+++ modules/@ergonode/attributes/src/pages/attribute-groups/index.vue
@@ -34,7 +34,7 @@
 <script>
 import CreateAttributeGroupButton from '@Attributes/components/Buttons/CreateAttributeGroupButton';
 import PRIVILEGES from '@Attributes/config/privileges';
-import beforeLeavePageMixin from '@Core/mixins/page/beforeLeavePageMixin';
+import beforeRouteLeaveMixin from '@Core/mixins/route/beforeRouteLeaveMixin';
 import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
 import Page from '@UI/components/Layout/Page';
 import HorizontalRoutingTabBar from '@UI/components/TabBar/Routing/HorizontalRoutingTabBar';
@@ -49,7 +49,7 @@ export default {
         HorizontalRoutingTabBar,
     },
     mixins: [
-        beforeLeavePageMixin,
+        beforeRouteLeaveMixin,
         asyncTabsMixin,
     ],
     computed: {
