--- modules/@ergonode/categories/src/components/Pages/CategoryPage.vue
@@ -1,46 +1,0 @@
-/*
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE for license details.
- */
-<template>
-    <Page>
-        <TitleBar
-            :title="title"
-            :is-read-only="isReadOnly">
-            <template #prependHeader>
-                <NavigateBackFab />
-            </template>
-            <template #mainAction>
-                <RemoveCategoryButton />
-            </template>
-        </TitleBar>
-        <HorizontalRoutingTabBar
-            v-if="asyncTabs"
-            :items="asyncTabs"
-            :change-values="changeValues"
-            :errors="errors" />
-    </Page>
-</template>
-
-<script>
-import RemoveCategoryButton from '@Categories/components/Buttons/RemoveCategoryButton';
-import PRIVILEGES from '@Categories/config/privileges';
-import editPageMixin from '@Core/mixins/page/editPageMixin';
-import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
-
-export default {
-    name: 'CategoryPage',
-    components: {
-        RemoveCategoryButton,
-    },
-    mixins: [
-        editPageMixin,
-        asyncTabsMixin,
-    ],
-    computed: {
-        isReadOnly() {
-            return this.$isReadOnly(PRIVILEGES.CATEGORY.namespace);
-        },
-    },
-};
-</script>
