--- modules/@ergonode/category-trees/src/components/Pages/CategoryTreePage.vue
@@ -1,46 +1,0 @@
-/*
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE for license details.
- */
-<template>
-    <Page>
-        <TitleBar
-            :title="title"
-            :is-read-only="isReadOnly">
-            <template #prependHeader>
-                <NavigateBackFab />
-            </template>
-            <template #mainAction>
-                <RemoveCategoryTreeButton />
-            </template>
-        </TitleBar>
-        <HorizontalRoutingTabBar
-            v-if="asyncTabs"
-            :items="asyncTabs"
-            :change-values="changeValues"
-            :errors="errors" />
-    </Page>
-</template>
-
-<script>
-import editPageMixin from '@Core/mixins/page/editPageMixin';
-import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
-import RemoveCategoryTreeButton from '@Trees/components/Buttons/RemoveCategoryTreeButton';
-import PRIVILEGES from '@Trees/config/privileges';
-
-export default {
-    name: 'CategoryTreePage',
-    components: {
-        RemoveCategoryTreeButton,
-    },
-    mixins: [
-        editPageMixin,
-        asyncTabsMixin,
-    ],
-    computed: {
-        isReadOnly() {
-            return this.$isReadOnly(PRIVILEGES.CATEGORY_TREE.namespace);
-        },
-    },
-};
-</script>
