+++ modules/@ergonode/category-trees/src/pages/category-trees/_tree/_id.vue
@@ -3,15 +3,37 @@
  * See LICENSE for license details.
  */
 <template>
-    <CategoryTreePage :title="code" />
+    <Page>
+        <TitleBar
+            :title="code"
+            :is-read-only="isReadOnly">
+            <template #prependHeader>
+                <NavigateBackFab :previous-route="previousRoute" />
+            </template>
+            <template #mainAction>
+                <RemoveCategoryTreeButton />
+            </template>
+        </TitleBar>
+        <HorizontalRoutingTabBar
+            v-if="asyncTabs"
+            :items="asyncTabs"
+            :change-values="changeValues"
+            :errors="errors" />
+    </Page>
 </template>
 
 <script>
 import {
     ALERT_TYPE,
 } from '@Core/defaults/alerts';
-import beforeLeavePageMixin from '@Core/mixins/page/beforeLeavePageMixin';
-import CategoryTreePage from '@Trees/components/Pages/CategoryTreePage';
+import beforeRouteEnterMixin from '@Core/mixins/route/beforeRouteEnterMixin';
+import beforeRouteLeaveMixin from '@Core/mixins/route/beforeRouteLeaveMixin';
+import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
+import RemoveCategoryTreeButton from '@Trees/components/Buttons/RemoveCategoryTreeButton';
+import PRIVILEGES from '@Trees/config/privileges';
+import Page from '@UI/components/Layout/Page';
+import HorizontalRoutingTabBar from '@UI/components/TabBar/Routing/HorizontalRoutingTabBar';
+import TitleBar from '@UI/components/TitleBar/TitleBar';
 import {
     mapActions,
     mapState,
@@ -20,10 +42,15 @@ import {
 export default {
     name: 'EditCategoryTree',
     components: {
-        CategoryTreePage,
+        Page,
+        TitleBar,
+        HorizontalRoutingTabBar,
+        RemoveCategoryTreeButton,
     },
     mixins: [
-        beforeLeavePageMixin,
+        asyncTabsMixin,
+        beforeRouteEnterMixin,
+        beforeRouteLeaveMixin,
     ],
     validate({
         params,
@@ -49,6 +76,9 @@ export default {
         ...mapState('categoryTree', [
             'code',
         ]),
+        isReadOnly() {
+            return this.$isReadOnly(PRIVILEGES.CATEGORY_TREE.namespace);
+        },
     },
     beforeDestroy() {
         this.__clearGridDesignerStorage();
