--- modules/@ergonode/channels/src/components/Pages/ChannelPage.vue
@@ -1,49 +1,0 @@
-/*
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE for license details.
- */
-<template>
-    <Page>
-        <TitleBar
-            :title="title"
-            :is-read-only="isReadOnly">
-            <template #prependHeader>
-                <NavigateBackFab />
-            </template>
-            <template #mainAction>
-                <RemoveChannelButton />
-                <CreateExportButton />
-            </template>
-        </TitleBar>
-        <HorizontalRoutingTabBar
-            v-if="asyncTabs"
-            :items="asyncTabs"
-            :change-values="changeValues"
-            :errors="errors" />
-    </Page>
-</template>
-
-<script>
-import CreateExportButton from '@Channels/components/Buttons/CreateExportButton';
-import RemoveChannelButton from '@Channels/components/Buttons/RemoveChannelButton';
-import PRIVILEGES from '@Channels/config/privileges';
-import editPageMixin from '@Core/mixins/page/editPageMixin';
-import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
-
-export default {
-    name: 'ChannelPage',
-    components: {
-        CreateExportButton,
-        RemoveChannelButton,
-    },
-    mixins: [
-        editPageMixin,
-        asyncTabsMixin,
-    ],
-    computed: {
-        isReadOnly() {
-            return this.$isReadOnly(PRIVILEGES.CHANNEL.namespace);
-        },
-    },
-};
-</script>
