+++ modules/@ergonode/channels/src/pages/channels/index.vue
@@ -20,7 +20,7 @@
 <script>
 import CreateChannelButton from '@Channels/components/Buttons/CreateChannelButton';
 import PRIVILEGES from '@Channels/config/privileges';
-import beforeLeavePageMixin from '@Core/mixins/page/beforeLeavePageMixin';
+import beforeRouteLeaveMixin from '@Core/mixins/route/beforeRouteLeaveMixin';
 import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
 import Page from '@UI/components/Layout/Page';
 import HorizontalRoutingTabBar from '@UI/components/TabBar/Routing/HorizontalRoutingTabBar';
@@ -35,7 +35,7 @@ export default {
         HorizontalRoutingTabBar,
     },
     mixins: [
-        beforeLeavePageMixin,
+        beforeRouteLeaveMixin,
         asyncTabsMixin,
     ],
     computed: {
