--- modules/@ergonode/import/src/components/Pages/ImportProfilePage.vue
@@ -1,49 +1,0 @@
-/*
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE for license details.
- */
-<template>
-    <Page>
-        <TitleBar
-            :title="title"
-            :is-read-only="isReadOnly">
-            <template #prependHeader>
-                <NavigateBackFab />
-            </template>
-            <template #mainAction>
-                <RemoveImportButton />
-                <CreateImportButton />
-            </template>
-        </TitleBar>
-        <HorizontalRoutingTabBar
-            v-if="asyncTabs"
-            :items="asyncTabs"
-            :change-values="changeValues"
-            :errors="errors" />
-    </Page>
-</template>
-
-<script>
-import editPageMixin from '@Core/mixins/page/editPageMixin';
-import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
-import CreateImportButton from '@Import/components/Buttons/CreateImportButton';
-import RemoveImportButton from '@Import/components/Buttons/RemoveImportButton';
-import PRIVILEGES from '@Import/config/privileges';
-
-export default {
-    name: 'ImportProfilePage',
-    components: {
-        CreateImportButton,
-        RemoveImportButton,
-    },
-    mixins: [
-        editPageMixin,
-        asyncTabsMixin,
-    ],
-    computed: {
-        isReadOnly() {
-            return this.$isReadOnly(PRIVILEGES.IMPORT.namespace);
-        },
-    },
-};
-</script>
