--- modules/@ergonode/product-templates/src/components/Pages/TemplatePage.vue
@@ -1,46 +1,0 @@
-/*
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE for license details.
- */
-<template>
-    <Page>
-        <TitleBar
-            :title="title"
-            :is-read-only="isReadOnly">
-            <template #prependHeader>
-                <NavigateBackFab />
-            </template>
-            <template #mainAction>
-                <RemoveProductTemplateButton />
-            </template>
-        </TitleBar>
-        <HorizontalRoutingTabBar
-            v-if="asyncTabs"
-            :items="asyncTabs"
-            :change-values="changeValues"
-            :errors="errors" />
-    </Page>
-</template>
-
-<script>
-import editPageMixin from '@Core/mixins/page/editPageMixin';
-import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
-import RemoveProductTemplateButton from '@Templates/components/Buttons/RemoveProductTemplateButton';
-import PRIVILEGES from '@Templates/config/privileges';
-
-export default {
-    name: 'TemplatePage',
-    components: {
-        RemoveProductTemplateButton,
-    },
-    mixins: [
-        editPageMixin,
-        asyncTabsMixin,
-    ],
-    computed: {
-        isReadOnly() {
-            return this.$isReadOnly(PRIVILEGES.TEMPLATE_DESIGNER.namespace);
-        },
-    },
-};
-</script>
