--- modules/@ergonode/segments/src/components/Pages/SegmentPage.vue
@@ -1,46 +1,0 @@
-/*
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE for license details.
- */
-<template>
-    <Page>
-        <TitleBar
-            :title="title"
-            :is-read-only="isReadOnly">
-            <template #prependHeader>
-                <NavigateBackFab />
-            </template>
-            <template #mainAction>
-                <RemoveSegmentButton />
-            </template>
-        </TitleBar>
-        <HorizontalRoutingTabBar
-            v-if="asyncTabs"
-            :items="asyncTabs"
-            :change-values="changeValues"
-            :errors="errors" />
-    </Page>
-</template>
-
-<script>
-import editPageMixin from '@Core/mixins/page/editPageMixin';
-import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
-import RemoveSegmentButton from '@Segments/components/Buttons/RemoveSegmentButton';
-import PRIVILEGES from '@Segments/config/privileges';
-
-export default {
-    name: 'SegmentPage',
-    components: {
-        RemoveSegmentButton,
-    },
-    mixins: [
-        editPageMixin,
-        asyncTabsMixin,
-    ],
-    computed: {
-        isReadOnly() {
-            return this.$isReadOnly(PRIVILEGES.SEGMENT.namespace);
-        },
-    },
-};
-</script>
