--- modules/@ergonode/status-transitions/src/components/Pages/TransitionPage.vue
@@ -1,46 +1,0 @@
-/*
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE for license details.
- */
-<template>
-    <Page>
-        <TitleBar
-            :title="title"
-            :is-read-only="isReadOnly">
-            <template #prependHeader>
-                <NavigateBackFab />
-            </template>
-            <template #mainAction>
-                <RemoveStatusTransitionButton />
-            </template>
-        </TitleBar>
-        <HorizontalRoutingTabBar
-            v-if="asyncTabs"
-            :items="asyncTabs"
-            :change-values="changeValues"
-            :errors="errors" />
-    </Page>
-</template>
-
-<script>
-import editPageMixin from '@Core/mixins/page/editPageMixin';
-import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
-import RemoveStatusTransitionButton from '@Transitions/components/Buttons/RemoveStatusTransitionButton';
-import PRIVILEGES from '@Transitions/config/privileges';
-
-export default {
-    name: 'TransitionPage',
-    components: {
-        RemoveStatusTransitionButton,
-    },
-    mixins: [
-        editPageMixin,
-        asyncTabsMixin,
-    ],
-    computed: {
-        isReadOnly() {
-            return this.$isReadOnly(PRIVILEGES.WORKFLOW.namespace);
-        },
-    },
-};
-</script>
