+++ modules/@ergonode/ui/src/components/AdvancedFilters/AdvancedFilterShowOnly.vue
@@ -25,7 +25,7 @@ export default {
          */
         value: {
             type: Boolean,
-            required: true,
+            default: false,
         },
     },
     methods: {
